# Optimizes initial states using SummedPotential