--- a/tests/test_single_topology.py
+++ b/tests/test_single_topology.py
@@ -463,6 +463,7 @@ def test_nonbonded_intra_split(precision, rtol, atol, use_tiny_mol):
     ffs = load_split_forcefields()
     solvent_sys, solvent_conf, solvent_box, solvent_top = build_water_system(4.0, ffs.ref.water_ff)
     solvent_bps, _ = openmm_deserializer.deserialize_system(solvent_sys, cutoff=1.2)
+    solv_sys = convert_bps_into_system(solvent_bps)
 
     def get_vacuum_solvent_u_grads(ff, lamb):
         st = SingleTopology(mol_a, mol_b, core, ff)
@@ -471,13 +472,13 @@ def test_nonbonded_intra_split(precision, rtol, atol, use_tiny_mol):
 
         vacuum_system = st.setup_intermediate_state(lamb)
         vacuum_potentials = vacuum_system.get_U_fns()
-        val_and_grad_fn = minimizer.get_val_and_grad_fn(vacuum_potentials, solvent_box)
+        val_and_grad_fn = minimizer.get_val_and_grad_fn(vacuum_potentials, solvent_box, precision=precision)
         vacuum_u, vacuum_grad = val_and_grad_fn(ligand_conf)
 
-        solvent_system = st.combine_with_host(convert_bps_into_system(solvent_bps), lamb, solvent_conf.shape[0])
+        solvent_system = st.combine_with_host(solv_sys, lamb, solvent_conf.shape[0])
         solvent_potentials = solvent_system.get_U_fns()
-        val_and_grad_fn = minimizer.get_val_and_grad_fn(solvent_potentials, solvent_box)
-        solvent_u, solvent_grad = val_and_grad_fn(combined_conf)
+        solv_val_and_grad_fn = minimizer.get_val_and_grad_fn(solvent_potentials, solvent_box, precision=precision)
+        solvent_u, solvent_grad = solv_val_and_grad_fn(combined_conf)
         return vacuum_grad, vacuum_u, solvent_grad, solvent_u
 
     n_lambdas = 3
@@ -609,7 +610,7 @@ def test_combine_with_host_split(precision, rtol, atol):
 
         combined_system = st.combine_with_host(host_system, lamb, num_water_atoms)
         potentials = combined_system.get_U_fns()
-        u, grad = minimizer.get_val_and_grad_fn(potentials, box)(combined_conf)
+        u, grad = minimizer.get_val_and_grad_fn(potentials, box, precision=precision)(combined_conf)
         return grad, u
 
     def compute_new_grad_u(ff: Forcefield, precision, x0, box, lamb, num_water_atoms, host_bps):
@@ -621,7 +622,7 @@ def test_combine_with_host_split(precision, rtol, atol):
 
         combined_system = st.combine_with_host(host_system, lamb, num_water_atoms)
         potentials = combined_system.get_U_fns()
-        u, grad = minimizer.get_val_and_grad_fn(potentials, box)(combined_conf)
+        u, grad = minimizer.get_val_and_grad_fn(potentials, box, precision=precision)(combined_conf)
         return grad, u
 
     def compute_intra_grad_u(ff: Forcefield, precision, x0, box, lamb, num_water_atoms, num_host_atoms):
@@ -630,7 +631,7 @@ def test_combine_with_host_split(precision, rtol, atol):
 
         vacuum_system = st.setup_intermediate_state(lamb)
         potentials = vacuum_system.get_U_fns()
-        u, grad = minimizer.get_val_and_grad_fn(potentials, box)(ligand_conf)
+        u, grad = minimizer.get_val_and_grad_fn(potentials, box, precision=precision)(ligand_conf)
 
         # Pad g so it's the same shape as the others
         grad_padded = np.concatenate([np.zeros((num_host_atoms, 3)), grad])