* About 10% faster for some tests (locally, does not seem to be the case in CI)

Running the potentials previously shows to take 4ms vs 1ms using nsys, but doesn't seem to translate to the tests. Unsure whether or not we should merge this, but opening up for consideration.
# Current
Takes 4ms to run the bonded potentials plus the nonbonded potentials
![old_version](https://github.com/proteneer/timemachine/assets/5840832/c0ced096-478a-4a08-90e0-6a259d91f5b7)

# PR
Takes ~1ms to run all of the potentials
![new](https://github.com/proteneer/timemachine/assets/5840832/ce620956-a21a-4fdf-8aa6-e88955726cf8)